-Refactor backend code; -Removed unnecessary folders and files;
