--- webui/models/completion_task.py
@@ -1,6 +1,0 @@
-from pydantic import BaseModel
-
-
-class TaskCreator(BaseModel):
-    prompt: str
-    mode: str
